+++ dbms/src/Functions/bitCount.cpp
@@ -26,7 +26,7 @@ struct BitCountImpl
         if constexpr (std::is_same_v<A, Int8>)
             return __builtin_popcount(static_cast<UInt8>(a));
         else
-            return __builtin_popcountll(ext::bit_cast<unsigned long long>(a));
+            return __builtin_popcountll(ext::bit_cast<uint64_t>(a));
     }
 
 #if USE_EMBEDDED_COMPILER
